refactor: extract build orchestration into service/ layer
