+++ builders/server/main.py
@@ -3,30 +3,12 @@ import logging
 import pandas as pd
 from fastapi import FastAPI, HTTPException, Query
 
-import db.datasets
-from runtime import config, loader, runner, validator
+from service.builder import build_dataset
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 app = FastAPI()
-
-GRANULARITY_MAP = {
-    "1s": "s",
-    "1m": "min",
-    "1h": "h",
-    "1d": "D",
-}
-
-
-def generate_timestamps(
-    start: pd.Timestamp, end: pd.Timestamp, granularity: str
-) -> list[pd.Timestamp]:
-    """Generate all timestamps in [start, end] for the given granularity."""
-    freq = GRANULARITY_MAP.get(granularity)
-    if freq is None:
-        raise ValueError(f"Unsupported granularity: {granularity}")
-    return list(pd.date_range(start=start, end=end, freq=freq))
 
 
 @app.post("/build/{dataset_name}/{dataset_version}")
@@ -46,72 +28,9 @@ def build(
         ) from exc
 
     try:
-        _build_dataset(dataset_name, dataset_version, start_ts, end_ts)
+        build_dataset(dataset_name, dataset_version, start_ts, end_ts)
     except Exception as e:
         logger.exception(f"Build failed for {dataset_name}/{dataset_version}")
         raise HTTPException(status_code=500, detail=str(e)) from e
 
     return {"status": "ok"}
-
-
-def _build_dataset(
-    dataset_name: str,
-    dataset_version: str,
-    start: pd.Timestamp,
-    end: pd.Timestamp,
-) -> None:
-    """Core build logic: resolve dependencies, build missing timestamps, insert rows."""
-    cfg = config.load_config(dataset_name, dataset_version)
-    dependencies = cfg.get("dependencies", {})
-    schema = cfg.get("schema", {})
-    granularity = cfg.get("granularity", "1d")
-
-    # Recursively build dependencies first
-    for dep_name, dep_version in dependencies.items():
-        _build_dataset(dep_name, dep_version, start, end)
-
-    # Determine which timestamps are missing
-    all_timestamps = generate_timestamps(start, end, granularity)
-    existing = set(
-        db.datasets.get_existing_timestamps(dataset_name, dataset_version, start, end)
-    )
-    missing = [ts for ts in all_timestamps if ts not in existing]
-
-    if not missing:
-        logger.info(
-            f"{dataset_name}/{dataset_version}: all timestamps present, skipping"
-        )
-        return
-
-    logger.info(
-        f"{dataset_name}/{dataset_version}: building {len(missing)} missing timestamps"
-    )
-
-    # Load the builder function
-    build_fn = loader.load_builder(dataset_name, dataset_version)
-
-    # Build each missing timestamp
-    rows = []
-    for ts in missing:
-        # Fetch dependency data for this timestamp
-        dep_data = {}
-        for dep_name, dep_version in dependencies.items():
-            dep_rows = db.datasets.get_rows(dep_name, dep_version, [ts])
-            if ts not in dep_rows:
-                raise RuntimeError(
-                    f"Dependency '{dep_name}/{dep_version}' "
-                    f"missing data for timestamp {ts}"
-                )
-            dep_data[dep_name] = dep_rows[ts]
-
-        # Run builder in subprocess
-        result = runner.run_builder(build_fn, dep_data, ts)
-
-        # Validate output against schema
-        validator.validate(result, schema)
-
-        rows.append((ts, result))
-
-    # Bulk insert all rows
-    db.datasets.insert_rows(dataset_name, dataset_version, rows)
-    logger.info(f"{dataset_name}/{dataset_version}: inserted {len(rows)} rows")
